chatbot-workflow/requirements.txt
langgraph>=0.1.0
langchain>=0.3.0
python-dotenv>=1.0.0
langchain-google-genai>=2.0.0
langchain-openai>=0.1.0
streamlit>=1.20.0